pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rocotoviewer"
version = "0.1.0"
description = "A powerful viewer for Rocoto workflow management systems"
readme = "README.md"
license = "MIT"
authors = [
    {name = "RocotoViewer Team", email = "rocotoviewer@example.com"}
]
maintainers = [
    {name = "RocotoViewer Team", email = "rocotoviewer@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Monitoring",
    "Framework :: Textual"
]
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "rich>=12.0.0",
    "textual>=0.40.0",
    "watchdog>=3.0.0",
    "colorama>=0.4.0; platform_system=='Windows'"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=2.0",
    "flake8>=5.0",
    "mypy>=1.0",
    "pre-commit>=3.0"
]
docs = [
    "mkdocs>=1.4.0",
    "mkdocs-material>=9.0"
]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0"
]

[project.scripts]
rocotoviewer = "rocotoviewer.cli:main"

[project.urls]
Homepage = "https://github.com/rocotoviewer/rocotoviewer"
Repository = "https://github.com/rocotoviewer/rocotoviewer"
Documentation = "https://rocotoviewer.readthedocs.io/"
Issues = "https://github.com/rocotoviewer/rocotoviewer/issues"
Download = "https://pypi.org/project/rocotoviewer/#files"

[tool.setuptools.packages.find]
where = ["."]
include = ["rocotoviewer*"]

[tool.setuptools.package-data]
rocotoviewer = ["examples/*.xml", "examples/*.yaml", "docs/*.md"]

[tool.setuptools.exclude-package-data]
rocotoviewer = ["tests*", "docs/*_draft.md"]